Universal Render Pipeline, Attempt 1 at Ski Controls
Changed rendering pipeline to URP so I could use shadergraph. Trying to make glittery snow, but it's not working great on mountain yet.
Switched from a CharacterController to a Rigidbody and CapsuleCollider. This lets us use physics much more easily for skiing controls. I tried adding skiing controls also but there's a lot of tweaking to be done